backend/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
langchain-community
pinecone-client
supabase
python-dotenv
pydantic
langchain-pinecone
pypdf
python-docx
unstructured
